requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pdfplumber==0.11.4
pymupdf==1.24.8
playwright==1.48.0
google-genai==0.8.0
